Small fix, increasing compiler strickness for tests

+++ inc/nix/fs.hpp
@@ -90,7 +90,7 @@ struct fs
     {
         if (path.length() == 0) { return Result::OK; }
 
-        for (int i = 0; i < path.size(); ++i)
+        for (size_t i = 0; i < path.size(); ++i)
         {
             if (path[i] == '/' || i == path.size() - 1)
             {
